// src/components/pages/auth/components/types.tsx
export interface LoginFormTypes {
  username?: string;
  password?: string;
  remember?: boolean;
  providerName?: string;
  redirectPath?: string;
}